[ENTESB-18353] Remove log4j references by upgrading to docker-java 3.…

+++ docker/docker/src/main/java/org/arquillian/cube/docker/impl/docker/DockerClientExecutor.java
@@ -394,7 +394,7 @@ public class DockerClientExecutor {
             }
 
             if (containerConfiguration.getCpuQuota() != null) {
-                createContainerCmd.getHostConfig().withCpuQuota(containerConfiguration.getCpuQuota());
+                createContainerCmd.getHostConfig().withCpuQuota((long) containerConfiguration.getCpuQuota());
             }
 
             if (containerConfiguration.getAttachStdin() != null) {
